+++ spring-cloud-tencent-rpc-enhancement/src/main/java/com/tencent/cloud/rpc/enhancement/resttemplate/EnhancedRestTemplateReporter.java
@@ -58,7 +58,14 @@ import static com.tencent.cloud.common.constant.ContextConstant.UTF_8;
  */
 public class EnhancedRestTemplateReporter extends AbstractPolarisReporterAdapter implements ResponseErrorHandler, ApplicationContextAware {
 
-	static final String HEADER_HAS_ERROR = "X-SCT-Has-Error";
+	/**
+	 * Polaris-CircuitBreaker-Fallback header flag.
+	 */
+	public static final String POLARIS_CIRCUIT_BREAKER_FALLBACK_HEADER = "X-SCT-Polaris-CircuitBreaker-Fallback";
+	/**
+	 * response has error header flag, since EnhancedRestTemplateReporter#hasError always return true.
+	 */
+	public static final String HEADER_HAS_ERROR = "X-SCT-Has-Error";
 	private static final Logger LOGGER = LoggerFactory.getLogger(EnhancedRestTemplateReporter.class);
 	private final ConsumerAPI consumerAPI;
 	private ResponseErrorHandler delegateHandler;
@@ -119,6 +126,9 @@ public class EnhancedRestTemplateReporter extends AbstractPolarisReporterAdapter
 	}
 
 	private void reportResult(URI url, ClientHttpResponse response) {
+		if (Boolean.parseBoolean(response.getHeaders().getFirst(POLARIS_CIRCUIT_BREAKER_FALLBACK_HEADER))) {
+			return;
+		}
 		try {
 			ServiceCallResult resultRequest = createServiceCallResult(url, response);
 			Map<String, String> loadBalancerContext = MetadataContextHolder.get().getLoadbalancerMetadata();
@@ -194,10 +204,8 @@ public class EnhancedRestTemplateReporter extends AbstractPolarisReporterAdapter
 	}
 
 	private void clear(ClientHttpResponse response) {
-		if (!response.getHeaders().containsKey(HEADER_HAS_ERROR)) {
-			return;
-		}
 		response.getHeaders().remove(HEADER_HAS_ERROR);
+		response.getHeaders().remove(POLARIS_CIRCUIT_BREAKER_FALLBACK_HEADER);
 	}
 
 	private ServiceCallResult createServiceCallResult(URI uri, ClientHttpResponse response) throws IOException {